Redefine sampling from memory
Now the sampling from the replay memory that happens in optimize_model
is repeated until the sample has at least one non-final next state.
Q-Learning cannot work with samples that have only empty next states,
as it uses the estimate of the next state values to compute the updates.
This should solve an exception that was occasionally being thrown by
PyTorch.